Fix bogus brightness changes when using dbus mode
Changing brightness value via DBUS is async and does not instantly result in actually changing the value in the underlying file.
This can result in changing the value via DBUS, and then still reading the unchanged value from the file, and thinking that it's a new user value.
As such, return cached "desired" value until the file was actually written after changing it via DBUS.